Rename leastSq -> leastSquares
For consistency with other op names; see:
https://github.com/imagej/imagej-ops/pull/570#issuecomment-418785981
